order-service/requirements.txt
backoff==2.2.1
black==22.10.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
confluent-kafka==1.8.2
Deprecated==1.2.13
Flask==2.2.2
googleapis-common-protos==1.56.4
grpcio==1.49.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
opentelemetry-api==1.13.0
opentelemetry-distro==0.34b0
opentelemetry-exporter-otlp==1.13.0
opentelemetry-exporter-otlp-proto-grpc==1.13.0
opentelemetry-exporter-otlp-proto-http==1.13.0
opentelemetry-instrumentation==0.34b0
opentelemetry-instrumentation-confluent-kafka==0.34b0
opentelemetry-instrumentation-flask==0.34b0
opentelemetry-instrumentation-wsgi==0.34b0
opentelemetry-proto==1.13.0
opentelemetry-sdk==1.13.0
opentelemetry-semantic-conventions==0.34b0
opentelemetry-util-http==0.34b0
pathspec==0.10.1
platformdirs==2.5.2
protobuf==3.20.3
requests==2.28.1
six==1.16.0
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
